[ADD]基本成型

+++ util/num.go
@@ -9,6 +9,6 @@ func Int32Min(a, b int32) int32 {
 	return b
 }
 
-func Distance(x1, y1, x2, y2 float64) float64 {
-	return math.Sqrt(math.Pow(x1-x2, 2) + math.Pow(y1-y2, 2))
+func Distance(x1, y1, x2, y2 float32) float32 {
+	return float32(math.Sqrt(math.Pow(float64(x1-x2), 2) + math.Pow(float64(y1-y2), 2)))
 }
